"appuifwmodule.rsc" loading with dynamic UID
This is done so apps do not interfere with "appuifwmodule.rsc" file bundled with the Python S60 v1.4.5 distribution. Each app will have a copy of this file with EXE UID in the name. This is one of the required steps for making apps that are truly portable without depending on system level Python installation. Also, this modification allows getting the file from installation drive, making embedding the app to firmware possible.